+++ NostraEngine/src/source/nostraengine/core/NostraEngine.cpp
@@ -3,6 +3,7 @@
 
 #include "nostraengine/core/NostraEngine.hpp"
 #include "nostraengine/core/PluginManager.hpp"
+#include "nostraengine/core/ConfigurationManagement.hpp"
 
 namespace NOE::NOE_CORE{
 
@@ -49,7 +50,7 @@ namespace NOE::NOE_CORE{
 
 		if (!PluginManager::get().createPluginList())
 		{
-			//NOU_LOG_ERROR("Failed to create the plugin list.");
+			NOU_LOG_ERROR("Failed to create the plugin list.");
 			return Initializable::ExitCode::ERROR;
 		}
 
@@ -71,20 +72,23 @@ namespace NOE::NOE_CORE{
 			switch (result)
 			{
 			case Plugin::InitResult::SUCCESS:
-				NOU_LOG_INFO(NOU::NOU_DAT_ALG::String8("The initialization of the plugin \"") + plugin->getMetadata().getName()
-					+ "\" (ID: " + plugin->getMetadata().getID() + ") was successful.");
+				NOU_LOG_INFO(NOU::NOU_DAT_ALG::String8("The initialization of the plugin \"") 
+					+ plugin->getMetadata().getName() + "\" (ID: " + plugin->getMetadata().getID() 
+					+ ") was successful.");
 				m_preInitializedObjects++;
 				break;
 			case Plugin::InitResult::WARNING:
-				NOU_LOG_WARNING(NOU::NOU_DAT_ALG::String8("The initialization of the plugin \"") + plugin->getMetadata().getName()
-					+ "\" (ID: " + plugin->getMetadata().getID() + ") has finished with a warning.");
+				NOU_LOG_WARNING(NOU::NOU_DAT_ALG::String8("The initialization of the plugin \"") 
+					+ plugin->getMetadata().getName() + "\" (ID: " + plugin->getMetadata().getID() 
+					+ ") has finished with a warning.");
 
 				ret = Initializable::ExitCode::WARNING;
 				m_preInitializedObjects++;
 				break;
 			case Plugin::InitResult::FAILED:
-				NOU_LOG_FATAL(NOU::NOU_DAT_ALG::String8("The initialization of the plugin \"") + plugin->getMetadata().getName()
-					+ "\" (ID: " + plugin->getMetadata().getID() + ") has failed.");
+				NOU_LOG_FATAL(NOU::NOU_DAT_ALG::String8("The initialization of the plugin \"") 
+					+ plugin->getMetadata().getName() + "\" (ID: " + plugin->getMetadata().getID() 
+					+ ") has failed.");
 				return Initializable::ExitCode::ERROR;
 			}
 		}
@@ -209,7 +213,8 @@ namespace NOE::NOE_CORE{
 
 	void NostraEngine::renderMain()
 	{
-		m_window->update();
+		if(m_window != nullptr)
+			m_window->update();
 	}
 
 	NOU::boolean NostraEngine::addUpdatable(Updatable *updt)
@@ -263,24 +268,28 @@ namespace NOE::NOE_CORE{
 		NOU::NOU_CORE::Logger::get().pushLogger<NOU::NOU_CORE::ConsoleLogger>();
 		//NOU::NOU_CORE::Logger::get().pushLogger<NOU::NOU_CORE::FileLogger>();
 
-		//NOU_LOG_INFO(NOU::NOU_DAT_ALG::String8("NostraEngine Version ") + getVersion().rawStr());
+		//construct instance of ConfigurationManager	
+		NOE::NOE_CORE::ConfigurationManager::get();
+
+		NOU_LOG_INFO(NOU::NOU_DAT_ALG::String8("NostraEngine Version ") + 
+			getVersion().getMajor() + "." + getVersion().getMinor() + "." + getVersion().getPatch());
 
 		NOE::NOE_CORE::PluginManager::get().initialize();
 
 		if(preInitialize() == Initializable::ExitCode::ERROR)
 		{
-		//	NOU_LOG_ERROR("preInitialize(): An Error occurred during pre initialize.");
+			NOU_LOG_ERROR("An Error occurred during pre-initialization.");
 			m_runState = -1;
 		}
 		else if (initialize() == Initializable::ExitCode::ERROR)
 		{
 
-		//	NOU_LOG_ERROR("Initialize(): An Error occurred during initialize.");
+			NOU_LOG_ERROR("An Error occurred during initialization.");
 			m_runState = -1;
 
 		}else if (postInitialize() == Initializable::ExitCode::ERROR)
 		{
-		//	NOU_LOG_ERROR("postInitialize(): An Error occurred during post initialize.");
+			NOU_LOG_ERROR("An Error occurred during post-initialization.");
 			m_runState = -1;
 		}
 
@@ -288,12 +297,12 @@ namespace NOE::NOE_CORE{
 
 		if (terminate() == Initializable::ExitCode::ERROR)
 		{
-		//	NOU_LOG_ERROR("terminate(): An Error occurred during terminate.");
+			NOU_LOG_ERROR("An Error occurred during termination.");
 		}
 
 		if (postTerminate() == Initializable::ExitCode::ERROR)
 		{
-		//	NOU_LOG_ERROR("postTerminate(): An Error occurred during post terminate.");
+			NOU_LOG_ERROR("postTerminate(): An Error occurred during post-termination.");
 		}
 
 		NOE::NOE_CORE::PluginManager::get().terminate();
@@ -334,7 +343,8 @@ namespace NOE::NOE_CORE{
 	{
 		if (&instance == nullptr)
 		{
-		//	NOU_PUSH_ERROR(NOU::NOU_CORE::getErrorHandler(), NOU::NOU_CORE::ErrorCodes::INVALID_STATE, "Cannot set the instance to a nullptr");
+			NOU_PUSH_ERROR(NOU::NOU_CORE::getErrorHandler(), NOU::NOU_CORE::ErrorCodes::INVALID_STATE, 
+				"Cannot set the instance to a nullptr");
 		}
 		else
 		{