+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Features
 - Added a plugin system
+- Added a resource management system
+- Added a configuration management system
+- Added the basic initialize-loop-terminate infrastructure
+- Added a basic SQLite database access interface
+- Added the plugin GLFWWindow
+- Added the plugin IniConfiguration
 
 ### Improvements
 - None